requirements.txt
# FastAPI and dependencies
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.7
pydantic==2.5.3
pydantic-settings==2.1.0
starlette==0.36.3

# Document processing
python-docx==1.1.0
PyMuPDF==1.23.21  # For PDF processing

# Data processing
pandas==2.1.4
openpyxl==3.1.2  # For Excel file generation

# OpenAI integration
openai==1.55.3
# Utilities
python-dotenv==1.0.0